exam_system/requirements.txt
flask
psycopg2-binary
flask-bcrypt
python-dotenv
gunicorn
